chore(backoffice-v2): resolved pr comment
