+++ apps/backoffice-v2/src/lib/blocks/variants/WebsiteMonitoringBlocks/hooks/useWebsiteMonitoringReportBlock/hooks/useWebsiteMonitoringBusinessReportTab/useWebsiteMonitoringBusinessReportTab.tsx
@@ -1,0 +1,179 @@
+import { TBusinessReport } from '@/domains/business-reports/fetchers';
+import { createReportAdapter } from '@/domains/business-reports/create-report-adapter/create-report-adapter';
+import React, { ReactNode, useCallback, useMemo } from 'react';
+import { WebsitesCompany } from '@/domains/business-reports/components/WebsitesCompany/WebsitesCompany';
+import { WebsiteLineOfBusiness } from '@/domains/business-reports/components/WebsiteLineOfBusiness/WebsiteLineOfBusiness';
+import { WebsiteCredibility } from '@/domains/business-reports/components/WebsiteCredibility/WebsiteCredibility';
+import { EcosystemAndTransactions } from '@/domains/business-reports/components/EcosystemAndTransactions/EcosystemAndTransactions';
+import { AdsAndSocialMedia } from '@/domains/business-reports/components/AdsAndSocialMedia/AdsAndSocialMedia';
+import { useSearchParamsByEntity } from '@/common/hooks/useSearchParamsByEntity/useSearchParamsByEntity';
+import { useLocation } from 'react-router-dom';
+
+export const useWebsiteMonitoringBusinessReportTab = ({
+  businessReport,
+}: {
+  businessReport: TBusinessReport;
+}) => {
+  const adapter = createReportAdapter({
+    reportVersion: businessReport?.report?.version,
+  });
+  const {
+    websitesCompanyAnalysis,
+    websiteCredibilityAnalysis,
+    adsAndSocialMediaAnalysis,
+    adsAndSocialMediaPresence,
+    websiteLineOfBusinessAnalysis,
+    ecosystemAndTransactionsAnalysis,
+    summary,
+    riskScore,
+    recommendations,
+    riskLevels,
+    companyReputationAnalysis,
+    relatedAdsSummary,
+    lineOfBusinessDescription,
+    onlineReputationAnalysis,
+    pricingAnalysis,
+    websiteStructureAndContentEvaluation,
+    trafficAnalysis,
+    ecosystemAndTransactionsMatches,
+    adsImages,
+    relatedAdsImages,
+  } = adapter(businessReport ?? {});
+  const tabs = useMemo(
+    () =>
+      [
+        {
+          label: "Website's Company",
+          value: 'websitesCompany',
+          content: (
+            <WebsitesCompany
+              companyReputationAnalysis={companyReputationAnalysis ?? []}
+              violations={websitesCompanyAnalysis ?? []}
+            />
+          ),
+        },
+        {
+          label: 'Website Line of Business',
+          value: 'websiteLineOfBusiness',
+          content: (
+            <WebsiteLineOfBusiness
+              violations={websiteLineOfBusinessAnalysis ?? []}
+              summary={lineOfBusinessDescription}
+            />
+          ),
+        },
+        {
+          label: 'Website Credibility',
+          value: 'websiteCredibility',
+          content: (
+            <WebsiteCredibility
+              violations={websiteCredibilityAnalysis ?? []}
+              onlineReputationAnalysis={onlineReputationAnalysis ?? []}
+              pricingAnalysis={pricingAnalysis}
+              websiteStructureAndContentEvaluation={websiteStructureAndContentEvaluation}
+              trafficAnalysis={trafficAnalysis}
+            />
+          ),
+        },
+        {
+          label: 'Ecosystem and Transactions',
+          value: 'ecosystemAndTransactions',
+          content: (
+            <EcosystemAndTransactions
+              violations={ecosystemAndTransactionsAnalysis ?? []}
+              matches={ecosystemAndTransactionsMatches ?? []}
+            />
+          ),
+        },
+        {
+          label: 'Ads and Social Media',
+          value: 'adsAndSocialMedia',
+          content: (
+            <AdsAndSocialMedia
+              violations={adsAndSocialMediaAnalysis ?? []}
+              mediaPresence={adsAndSocialMediaPresence ?? []}
+              adsImages={adsImages}
+              relatedAdsImages={relatedAdsImages}
+              relatedAdsSummary={relatedAdsSummary}
+            />
+          ),
+        },
+      ] as const satisfies ReadonlyArray<{
+        label: string;
+        value: string;
+        content: ReactNode | ReactNode[];
+      }>,
+    [
+      adsAndSocialMediaAnalysis,
+      adsAndSocialMediaPresence,
+      adsImages,
+      companyReputationAnalysis,
+      ecosystemAndTransactionsAnalysis,
+      ecosystemAndTransactionsMatches,
+      lineOfBusinessDescription,
+      onlineReputationAnalysis,
+      pricingAnalysis,
+      relatedAdsImages,
+      relatedAdsSummary,
+      trafficAnalysis,
+      websiteCredibilityAnalysis,
+      websiteLineOfBusinessAnalysis,
+      websiteStructureAndContentEvaluation,
+      websitesCompanyAnalysis,
+    ],
+  );
+  const [{ activeTab }] = useSearchParamsByEntity();
+  const { search } = useLocation();
+  const updateActiveTab = useCallback(({ tab, search }: { tab: string; search: string }) => {
+    const searchParams = new URLSearchParams(search);
+
+    searchParams.set('activeTab', tab);
+
+    return searchParams.toString();
+  }, []);
+  const riskIndicators = [
+    {
+      title: "Website's Company Analysis",
+      search: updateActiveTab({ tab: 'websitesCompany', search }),
+      violations: websitesCompanyAnalysis ?? [],
+    },
+    {
+      title: 'Website Credibility Analysis',
+      search: updateActiveTab({ tab: 'websiteCredibility', search }),
+      violations: websiteCredibilityAnalysis,
+    },
+    {
+      title: 'Ads and Social Media Analysis',
+      search: updateActiveTab({ tab: 'adsAndSocialMedia', search }),
+      violations: adsAndSocialMediaAnalysis ?? [],
+    },
+    {
+      title: 'Website Line of Business Analysis',
+      search: updateActiveTab({ tab: 'websiteLineOfBusiness', search }),
+      violations: websiteLineOfBusinessAnalysis ?? [],
+    },
+    {
+      title: 'Ecosystem and Transactions Analysis View',
+      search: updateActiveTab({ tab: 'ecosystemAndTransactions', search }),
+      violations: ecosystemAndTransactionsAnalysis ?? [],
+    },
+  ] as const satisfies ReadonlyArray<{
+    title: string;
+    search: string;
+    violations: Array<{
+      label: string;
+      severity: string;
+    }>;
+  }>;
+
+  return {
+    activeTab,
+    riskIndicators,
+    riskLevels,
+    riskScore,
+    recommendations,
+    tabs,
+    summary,
+    updateActiveTab,
+  };
+};
